backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
alembic
aiogram
openai
langchain
langchain-openai
tiktoken
python-multipart
pytesseract
Pillow
cryptography
passlib[bcrypt]
httpx
pydantic[email]
